Tweaks and Fixes
- Changed many text areas to be 26pt font at half-scale
- Fixed a bug with ahlf-and-double and synchronizing the timers
- Changed industry scalar to use double instead of float
- Made BigRational parse floats as floats, rather than as doubles
- Chain mail armor + icon
- Items with Effectiveness > 1 now require fewer items on quests
(further distinguishing torches and lanterns).

+++ Assets/draco18s/artificer/items/Scalar.cs
@@ -3,15 +3,15 @@ using UnityEngine;
 
 namespace Assets.draco18s.artificer.items {
 	public class Scalar {
-		public static Scalar _0_RAW =		new Scalar(1.05f, new Color(0, 0, 255));
-		public static Scalar _1_REFINED =	new Scalar(1.07f, new Color(0, 225, 225));
-		public static Scalar _2_SIMPLE =	new Scalar(1.09f, new Color(0, 255, 0));
-		public static Scalar _3_COMPLEX =	new Scalar(1.10f, new Color(225, 255, 0));
-		public static Scalar _4_RARE =		new Scalar(1.11f, new Color(225, 0, 0));
+		public static Scalar _0_RAW =		new Scalar(1.05, new Color(0, 0, 255));
+		public static Scalar _1_REFINED =	new Scalar(1.07, new Color(0, 225, 225));
+		public static Scalar _2_SIMPLE =	new Scalar(1.09, new Color(0, 255, 0));
+		public static Scalar _3_COMPLEX =	new Scalar(1.10, new Color(225, 255, 0));
+		public static Scalar _4_RARE =		new Scalar(1.11, new Color(225, 0, 0));
 
-		public float amount;
+		public double amount;
 		public Color color;
-		private Scalar(float f, Color c) {
+		private Scalar(double f, Color c) {
 			amount = f;
 			color = c;
 		}
